BACKEND/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
holidays>=0.40
python-dotenv>=1.0.0
google-generativeai>=0.3.0